Wired in armor to characters
Initialized good characters with the starter armor set. Also, made a
small change in SetFactory.java, there was double code for Troll armor,
and none for Goblin. I assumed that was a copy/paste typo. Also, made an
extra check statement inside of ArmorSet.java. Without the check, a null
pointer exception will be thrown if the character has not been fully
equipped in all slots with armor pieces(like when they are first
initialized, and equipped piece by piece). There is probably a better
solution than the extra if statement, but it will at least be functional
for now.