Dungeon and GameCore updates
It is now possible to move around in the dungeon.  There are still some
issues when certain rooms are entered (most likely UVineRooms).  There
is a place in Command to add the ability to use items outside of combat
if someone wants to add that.  Will continue work tomorrow on debugging
Dungeon movement and the combat system.