Combat, balance, and loot
The combat system is now fully functional and can be tested via
"TestPartyCombat.java". It not to pretty in some areas that I wrote, but
it does work, and I think it is sufficient for our remaining project
time. I also, filled out the treasure chest class, and adjusted the
openChest() method in command. Lastly, I gave all hero's and increase in
attack power to speed up the combat for testing.