fixed print and help commands
Because of the way our game runs, whenever any command is issued the
current room is going to be entered again.  To fix this would require
substantial changes, so we will call it a feature that if you have to
ask for help, or to see the room again, you may be penalized by getting
attacked (depending on what room you are in, upvine rooms are considered
safe rooms if the boss there is already killed).